elisp: mioa701 barebox remote work
Enable barebox commander to be invoked remotely.
If mioa701-barebox-start is launched from a tramp buffer (remote file),
barebox commander will be on that remote host.

The commands now require the user to choose the target barebox
commander, unless there is only one (which defaults to this commander).

Moreover, centralize the kernel command line generation.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>